+++ SharpFloat/src/Helpers/UInt128Mul64ByShifted32To128.cs
@@ -43,9 +43,9 @@ namespace SharpFloat.Helpers {
         /// <param name="b"></param>
         /// <returns></returns>
         public static UInt128 Mul64ByShifted32To128(ulong a, uint b) {
-            ulong mid = (uint)a * b;
+            ulong mid = (ulong)(uint)a * b;
             var v0 = mid << 32;
-            var v64 = ((uint)(a >> 32)) * b + (mid >> 32);
+            var v64 = ((ulong)(uint)(a >> 32)) * b + (mid >> 32);
             return new UInt128(v64, v0);
         }
